+++ pharus/dynamic_api_gen.py
@@ -16,6 +16,7 @@ from base64 import b64decode
 from datetime import datetime
 import inspect
 import traceback
+import os
 try:
     from .component_interface_override import type_map
 except (ModuleNotFoundError, ImportError):
@@ -37,6 +38,25 @@ def {method_name}(jwt_payload: dict) -> dict:
         except Exception as e:
             return traceback.format_exc(), 500
 """
+    route_template_nologin = """
+
+@app.route('{route}', methods=['GET'])
+def {method_name}() -> dict:
+    if request.method in {{'GET'}}:
+        jwt_payload = dict(
+            databaseAddress=os.environ["PHARUS_HOST"],
+            username=os.environ["PHARUS_USER"],
+            password=os.environ["PHARUS_PASSWORD"],
+        )
+        try:
+            component_instance = type_map['{component_type}'](name='{component_name}',
+                                                              component_config={component},
+                                                              static_config={static_config},
+                                                              jwt_payload=jwt_payload)
+            return component_instance.{method_name_type}()
+        except Exception as e:
+            return traceback.format_exc(), 500
+"""
 
     pharus_root = f"{pkg_resources.get_distribution('pharus').module_path}/pharus"
     api_path = f"{pharus_root}/dynamic_api.py"
@@ -44,6 +64,9 @@ def {method_name}(jwt_payload: dict) -> dict:
     values_yaml = EnvYAML(Path(spec_path))
     with open(Path(api_path), "w") as f:
         f.write(header_template)
+        active_route_template = (
+            route_template if values_yaml["SciViz"]["auth"] else route_template_nologin
+        )
         if (
             "component_interface" in values_yaml["SciViz"]
             and "override" in values_yaml["SciViz"]["component_interface"]
@@ -74,10 +97,10 @@ def {method_name}(jwt_payload: dict) -> dict:
             for grid in page["grids"].values():
                 if grid["type"] == "dynamic":
                     f.write(
-                        route_template.format(
+                        (active_route_template).format(
                             route=grid["route"],
                             method_name=grid["route"].replace("/", ""),
-                            component_type="table",
+                            component_type="basicquery",
                             component_name="dynamicgrid",
                             component=json.dumps(grid),
                             static_config=static_config,
@@ -90,9 +113,12 @@ def {method_name}(jwt_payload: dict) -> dict:
                     if "component_templates" in grid
                     else grid["components"]
                 ).items():
-                    if re.match(r"^(table|metadata|plot|file).*$", comp["type"]):
+                    if re.match(
+                        r"^(table|metadata|plot|file|slider|dropdown-query).*$",
+                        comp["type"],
+                    ):
                         f.write(
-                            route_template.format(
+                            (active_route_template).format(
                                 route=comp["route"],
                                 method_name=comp["route"].replace("/", ""),
                                 component_type=comp["type"],
@@ -107,7 +133,7 @@ def {method_name}(jwt_payload: dict) -> dict:
                                 comp["type"]
                             ].attributes_route_format.format(route=comp["route"])
                             f.write(
-                                route_template.format(
+                                (active_route_template).format(
                                     route=attributes_route,
                                     method_name=attributes_route.replace("/", ""),
                                     component_type=comp["type"],