+++ pharus/interface.py
@@ -113,7 +113,7 @@ class _DJConnector:
         restriction: list = [],
         limit: int = 1000,
         page: int = 1,
-        order=["KEY ASC"],
+        order=None,
         fetch_blobs=False,
         fetch_args=[],
     ) -> tuple:
@@ -150,19 +150,25 @@ class _DJConnector:
             ]
         )
 
+        order_by = (
+            fetch_args.pop("order_by") if "order_by" in fetch_args else ["KEY ASC"]
+        )
+        order_by = order if order else order_by
+
+        limit = fetch_args.pop("limit") if "limit" in fetch_args else limit
+
         if fetch_blobs and not fetch_args:
             fetch_args = [*query.heading.attributes]
         elif not fetch_args:
             fetch_args = query.heading.non_blobs
         else:
             attributes = {k: v for k, v in attributes.items() if k in fetch_args}
-
         non_blobs_rows = query_restricted.fetch(
             *fetch_args,
             as_dict=True,
             limit=limit,
             offset=(page - 1) * limit,
-            order_by=order,
+            order_by=order_by,
         )
 
         # Buffer list to be return